Move v2 plugin work to v2 area
