+++ test/unit/plugin/v1/resource_test.rb
