+++ test/unit/plugin/v2/loader_test.rb
@@ -2,7 +2,7 @@
 
 require 'minitest/autorun'
 require 'minitest/test'
-require_relative '../../../lib/inspec/plugin'
+require_relative '../../../../lib/inspec/plugin/v2'
 
 require 'byebug'
 
@@ -12,7 +12,7 @@ class PluginLoaderTests < MiniTest::Test
 
   def reset_globals
     # These are effectively globals
-    Inspec::Plugin::Registry.instance.__reset
+    Inspec::Plugin::V2::Registry.instance.__reset
     ENV['HOME'] = @@orig_home
     ENV['INSPEC_CONFIG_DIR'] = nil
   end
@@ -20,7 +20,7 @@ class PluginLoaderTests < MiniTest::Test
   def setup
     reset_globals
 
-    repo_path = File.expand_path(File.join( __FILE__, '..', '..', '..', '..'))
+    repo_path = File.expand_path(File.join( __FILE__, '..', '..', '..', '..', '..'))
     mock_path = File.join(repo_path, 'test', 'unit', 'mock')
 
     @config_dir_path = File.join(mock_path, 'config_dirs')
@@ -36,22 +36,22 @@ class PluginLoaderTests < MiniTest::Test
   #====================================================================#
 
   def test_constructor_should_not_load_anything_automatically
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new
     assert_equal 0, reg.loaded_count, "\nRegistry load count"
   end
 
   def test_constructor_should_detect_bundled_plugins
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new
     @bundled_plugins.each do |bundled_plugin_name|
       assert reg.known_plugin?(bundled_plugin_name), "\n#{bundled_plugin_name} should be detected as a bundled plugin"
     end
   end
 
   def test_constructor_should_skip_bundles_when_option_is_set
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new(omit_bundles: true)
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new(omit_bundles: true)
     @bundled_plugins.each do |bundled_plugin_name|
       refute reg.known_plugin?(bundled_plugin_name), "\n#{bundled_plugin_name} should not be detected when omit_bundles is set"
     end
@@ -59,8 +59,8 @@ class PluginLoaderTests < MiniTest::Test
 
   def test_constructor_when_using_home_dir_detects_declared_plugins
     ENV['HOME'] = File.join(@config_dir_path, 'fakehome')
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new
     assert reg.known_plugin?(:'inspec-test-home-marker'), "\ninspec-test-home-marker should be detected as a plugin"
   end
 
@@ -70,8 +70,8 @@ class PluginLoaderTests < MiniTest::Test
 
   def test_constructor_when_the_plugin_config_is_absent_it_detects_bundled_plugins
     ENV['INSPEC_CONFIG_DIR'] = File.join(@config_dir_path, 'empty')
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new
     @bundled_plugins.each do |bundled_plugin_name|
       assert reg.known_plugin?(bundled_plugin_name), "\n#{bundled_plugin_name} should be detected as a bundled plugin"
     end
@@ -79,12 +79,12 @@ class PluginLoaderTests < MiniTest::Test
 
   def test_constuctor_when_the_plugin_config_is_corrupt_it_throws_an_exception
     ENV['INSPEC_CONFIG_DIR'] = File.join(@config_dir_path, 'corrupt')
-    assert_raises(Inspec::Plugin::ConfigError) { Inspec::Plugin::Loader.new }
+    assert_raises(Inspec::Plugin::V2::ConfigError) { Inspec::Plugin::V2::Loader.new }
   end
 
   def test_constuctor_when_the_plugin_config_is_a_bad_version_it_throws_an_exception
     ENV['INSPEC_CONFIG_DIR'] = File.join(@config_dir_path, 'bad_plugin_conf_version')
-    assert_raises(Inspec::Plugin::ConfigError) { Inspec::Plugin::Loader.new }
+    assert_raises(Inspec::Plugin::V2::ConfigError) { Inspec::Plugin::V2::Loader.new }
   end
 
   #====================================================================#
@@ -92,15 +92,15 @@ class PluginLoaderTests < MiniTest::Test
   #====================================================================#
 
   def test_load_no_plugins_should_load_no_plugins
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new(omit_bundles: true)
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new(omit_bundles: true)
     loader.load_all
     assert_equal 0, reg.loaded_count, "\nRegistry load count"
   end
 
   def test_load_only_bundled_plugins_should_load_bundled_plugins
-    reg = Inspec::Plugin::Registry.instance
-    loader = Inspec::Plugin::Loader.new
+    reg = Inspec::Plugin::V2::Registry.instance
+    loader = Inspec::Plugin::V2::Loader.new
     loader.load_all
     @bundled_plugins.each do |bundled_plugin_name|
       assert reg.loaded_plugin?(bundled_plugin_name), "\n#{bundled_plugin_name} should be loaded"
